pitrainer/requirements.txt
numpy==1.13.1
pigpio==1.35
pkg-resources==0.0.0
pysmbus==0.1
scrollphathd==1.0.1
